Stop nightly V2 head retraining to fix representation-leakage bias
The V2 temporal/confession heads were retrained nightly by train_pruning_model on
the stored encoder embeddings, with the new head replacing PROD when significantly
better. That accuracy is biased: the camembert body is label-fine-tuned on those
same sentences (train_encoder.py), so a head retrained/evaluated on the stored
embeddings is graded on inputs that already encode the label (double-dipping). It
also made the #42 encoder gate compare an honest joint head against a leaky nightly
incumbent, biasing it against promotion.

Option B (V2-only): V2 heads are produced only jointly with the encoder and
registered by promote_encoder; remove the nightly V2 retrain. V1 action (frozen
sentence-transformer, no leakage) is untouched.

- remove daily_train_pruning_model cron + delete the V2-only command
- train_classifier: ACTION-only, raise for V2; drop dead _train_encoder_head_classifier
- relocate get_test_size + MIN_DATASET_SIZE to utils/stat_utils so train_encoder
  (a workflow) can reuse the same stepped held-out size as V1 (527 -> 800)
- drop obsolete HEARTBEAT_TRAIN_PRUNING_URL; refresh CLAUDE.md + gate comment
