+++ ansible/prod/roles/cron/tasks/02_commands.yml
@@ -19,11 +19,6 @@
       hour: "0"
       minute: 40
       memory_max: "3200M"
-    - name: daily_train_pruning_model
-      command: "train_pruning_model --automatic"
-      hour: "0"
-      minute: 45
-      memory_max: "3200M"
     # V2 reclassify/outliers load the fine-tuned camembert-large encoder (~1.3 GB) to re-embed
     # stale sentences the night after an encoder promotion -> higher cap than the 3.2 GB default.
     - name: daily_reclassify_temporal
