+++ env.sample
@@ -54,7 +54,6 @@ export TROUVERUNEMESSE_USERNAME=secret
 export TROUVERUNEMESSE_PASSWORD=secret
 export HEARTBEAT_NIGHTLY_FETCHING_URL=secret
 export HEARTBEAT_CHECK_DATA_URL=secret
-export HEARTBEAT_TRAIN_PRUNING_URL=secret
 export HEARTBEAT_ANNUAIRE_URL=secret
 export DISCORD_NEW_REPORTS_URL=secret
 export DISCORD_NEW_IMAGES_URL=secret
